+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -5,10 +5,10 @@ package info.metadude.kotlin.library.engelsystem
 object GradlePlugins {
 
     private object Versions {
-        const val dokka = "1.7.20"
-        const val kotlin = "1.7.22"
-        const val ksp = "1.7.22-1.0.8"
-        const val versions = "0.44.0"
+        const val dokka = "1.8.20"
+        const val kotlin = "1.8.22"
+        const val ksp = "1.8.22-1.0.11"
+        const val versions = "0.48.0"
     }
 
     const val dokka = "org.jetbrains.dokka:dokka-gradle-plugin:${Versions.dokka}"
@@ -20,13 +20,13 @@ object GradlePlugins {
 object Libs {
 
     private object Versions {
-        const val junitJupiter = "5.9.1"
-        const val kotlinCoroutines = "1.6.4"
-        const val moshi = "1.14.0"
-        const val okhttp = "3.12.13"
-        const val retrofit = "2.6.4"
-        const val threetenbp = "1.6.5"
-        const val truth = "1.1.3"
+        const val junitJupiter = "5.10.0"
+        const val kotlinCoroutines = "1.7.3"
+        const val moshi = "1.15.0"
+        const val okhttp = "4.11.0"
+        const val retrofit = "2.9.0"
+        const val threetenbp = "1.6.8"
+        const val truth = "1.1.5"
     }
 
     const val junitJupiterApi = "org.junit.jupiter:junit-jupiter-api:${Versions.junitJupiter}"